FIX: Management of the SQL Numeric field type (to a Decimal type in CSharp)

+++ Test/Apps72.Dev.Data.Tests/Scott.cs
@@ -16,7 +16,7 @@ namespace Data.Tests
     {
         public int? SAL { get; set; }       // Not used, because there are a Salary property tagged [Column("SAL")]
         [Apps72.Dev.Data.Annotations.Column("sal")]
-        public int? Salary { get; set; }
+        public decimal? Salary { get; set; }
         [Apps72.Dev.Data.Annotations.Column("MGR")]
         public int? Manager { get; set; }
         public int? MGR { get; set; }       // Not used, because there are a Manager property tagged [Column("MGR")]
